// autogpt_platform/frontend/src/app/api/__generated__/models/userOnboarding.ts
/**
 * Generated by orval v7.13.0 🍺
 * Do not edit manually.
 * AutoGPT Agent Server
 * This server is used to execute agents that are created by the AutoGPT system.
 * OpenAPI spec version: 0.1
 */
import type { OnboardingStep } from "./onboardingStep";
import type { UserOnboardingUsageReason } from "./userOnboardingUsageReason";
import type { UserOnboardingOtherIntegrations } from "./userOnboardingOtherIntegrations";
import type { UserOnboardingSelectedStoreListingVersionId } from "./userOnboardingSelectedStoreListingVersionId";
import type { UserOnboardingAgentInput } from "./userOnboardingAgentInput";
import type { UserOnboardingOnboardingAgentExecutionId } from "./userOnboardingOnboardingAgentExecutionId";
import type { UserOnboardingLastRunAt } from "./userOnboardingLastRunAt";

export interface UserOnboarding {
  userId: string;
  completedSteps: OnboardingStep[];
  walletShown: boolean;
  notified: OnboardingStep[];
  rewardedFor: OnboardingStep[];
  usageReason: UserOnboardingUsageReason;
  integrations: string[];
  otherIntegrations: UserOnboardingOtherIntegrations;
  selectedStoreListingVersionId: UserOnboardingSelectedStoreListingVersionId;
  agentInput: UserOnboardingAgentInput;
  onboardingAgentExecutionId: UserOnboardingOnboardingAgentExecutionId;
  agentRuns: number;
  lastRunAt: UserOnboardingLastRunAt;
  consecutiveRunDays: number;
}